HydrostaticPressureKernel
