+++ src/coreComponents/constitutive/fluid/multifluid/CO2Brine/CO2BrineFluid.hpp
@@ -180,10 +180,7 @@ private:
   };
 
 protected:
-
   virtual void postInputInitialization() override;
-
-  virtual void initializePreSubGroups() override;
 
 private:
