+++ src/coreComponents/physicsSolvers/fluidFlow/CMakeLists.txt
@@ -98,6 +98,7 @@ set( fluidFlowSolvers_headers
      kernels/compositional/FluxComputeKernelBase.hpp
      kernels/compositional/GlobalComponentFractionKernel.hpp
      kernels/compositional/HydrostaticPressureKernel.hpp
+     kernels/compositional/HydrostaticPressureKernel_impl.hpp
      kernels/compositional/IHUPhaseFlux.hpp
      kernels/compositional/HU2PhaseFlux.hpp
      kernels/compositional/KernelLaunchSelectors.hpp
@@ -178,18 +179,16 @@ set( fluidFlowSolvers_sources
 
 # Remove the old generator-based hybrid flux/dirichlet kernels and switch to macro approach
 file( READ "${CMAKE_CURRENT_LIST_DIR}/kernelSpecs.json" kernelSpecs )
-set( kernelTemplateFileList "" )
-
-# Keep only the templates that are unrelated to hybrid flux/dirichlet
-list( APPEND kernelTemplateFileList
-      ReactiveCompositionalMultiphaseOBLKernels.cpp.template
-      CompositionalMultiphaseBase_FluidUpdateKernel.cpp.template
-      CompositionalMultiphaseFVM_IsothermalDirichletFluxComputeKernels.cpp.template
-      CompositionalMultiphaseFVM_ThermalDirichletFluxComputeKernels.cpp.template
-      CompositionalMultiphaseHybridFVMKernels_flux_grouped.cpp.template
-      CompositionalMultiphaseHybridFVMKernels_dirichlet_grouped.cpp.template
-      CompositionalMultiphaseHybridFVMKernels_bcprops.cpp.template
-      SinglePhaseHybridFVMKernels_grouped.cpp.template )
+set( kernelTemplateFileList 
+     ReactiveCompositionalMultiphaseOBLKernels.cpp.template
+     CompositionalMultiphaseBase_FluidUpdateKernel.cpp.template
+     CompositionalMultiphaseBase_HydrostaticPressureKernel.cpp.template
+     CompositionalMultiphaseFVM_IsothermalDirichletFluxComputeKernels.cpp.template
+     CompositionalMultiphaseFVM_ThermalDirichletFluxComputeKernels.cpp.template
+     CompositionalMultiphaseHybridFVMKernels_flux_grouped.cpp.template
+     CompositionalMultiphaseHybridFVMKernels_dirichlet_grouped.cpp.template
+     CompositionalMultiphaseHybridFVMKernels_bcprops.cpp.template
+     SinglePhaseHybridFVMKernels_grouped.cpp.template )
 
 foreach( kernelTemplateFile ${kernelTemplateFileList} )
   get_filename_component( jsonKey ${kernelTemplateFile} NAME_WE )
